content loader routes and controllers updated
